Add files via upload
I tried running what we have on here and I got some undefined reference errors. I edited main.cpp to include ZombieWar.cpp, I edited ZombieWar.cpp to include CharacterFactory.cpp, and I edited Soldier.h to look more like Lauren's Teacher.h so that way we wouldn't need a separate .cpp file for that one. After all of this, it compiles, it runs, but it hits a Segmentation Fault and the numbers it comes up with look a little wonky, if someone else wants to take another look and see what else needs to be fixed.